requirements.txt
requests
pandas
psycopg2-binary
sqlalchemy
streamlit
plotly
python-dotenv
schedule
